Day 18/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
httpx